+++ src/main/java/com/kt/dto/email/EmailResponse.java
@@ -6,14 +6,4 @@ public class EmailResponse {
             String message,
             String verificationCode
     ) {}
-
-    public record FindIdResponse(
-            boolean success,
-            String message
-    ) {
-        public static EmailResponse.FindIdResponse ok() {
-            return new EmailResponse.FindIdResponse(true, "아이디 안내 메일이 전송되었습니다.");
-        }
-    }
-
 }
